refresh .emacs config
added a way to install packages on startup (for a new install or
different machine)

took a stab at defining some commands to shell out to chunkc to
move around the emacs frame from within emacs (I have disabled my
shortcut manager skhd from working when emacs or iterm are in focus)

the above sort of works, but there is some latency and it doesn't seem
to work very consistently. I'll revisit this later.