fix emacs interaction with chunkc for window management
found

http://ergoemacs.org/emacs/elisp_call_shell_command.html which

led me to

https://www.gnu.org/software/emacs/manual/html_node/elisp/Asynchronous-Processes.html

and then swaped call-process-shell-command for
start-process-shell-command and got the snappiness that I was
expecting.

then bound it to free chords and it's working quite well. kinda
wondering if I ought to remap my skhd bindings to match emacs and live
with that.

for now I'm happy with this, but the different bindings is definitely
going to mess with me.