update: add oil shell to brewfile
tried to use oil as my shell, but it still seem too early for prime
time on mac os. Still, I want to keep it around to try it out as it
evolves. The idea of a new shell is interesting one (I suppose I could
also go back to fish)