remove org-roam
after a short time of really heavy use I kinda stopped using org-roam,
there were things I wanted to do with it, but ultimately it needed
more and more tweaking to get multiple separate components to
work. This is one of emacs strengths, but it can also lead to getting
things to not quite work the way you expect them in very minor ways,
not showstopping ways, but ways that after a while accrete.

cannot say for sure when I stopped using org-roam as much, but I did.

this commit removes org roam for a simpler note taking approach that
may be enhanced with some code that exists outside of emacs (maybe).

for now just going with some very simple org-capture set up.