feat: try out new window manager
i have used `yabai` for a long time, but ketp running into a few
things that i wasn't too happy with and i didn't like having to
also use `skhd` to manager keybindings despite it being pretty
configurable.

i briefly tried aerospace, but it does its own spaces management
and - while it does it for good reasons - it caused multiple issues
when i had to terminate it or it'd crash because it was a pixel
finding and dragging exercise to get them out.

i had starred `phoenix` a while back but never gave it a try so i
decided to try it out. `hammerspoon` was also an option (specially
since there's a couple of `PaperWM` implementations for it)... but
in the end i am more familiar with javascript :shrug:

ultimately i realized i don't want a fully tiling window manager,
but something that i can use to control layout that also provides
scriptability and configuration. and i'm starting with some very
basic window management behavior (like 3 operations lol).

the only downsides i have encountered so far are:
- cannot execute phoenix operations from cli, this can be useful.
- it's hard to see how things are executing, but i think this is a pilot error.